requirements.txt
pandas~=1.4.2
python-dateutil~=2.8.2
joblib~=1.1.0
tqdm~=4.64.0
scikit-learn~=1.1.1
numpy~=1.23.0
setuptools~=62.6.0
geocoder~=1.38.1
geopy~=2.2.0
flair~=0.11.3
dateparser~=1.1.0
word2number~=1.1
pybind11
fasttext-wheel
newspaper3k
psaw
torch>=1.10.2
certifi~=2020.6.20
urllib3~=1.25.8
websockets~=8.1
spacy~=3.2.0